PlayerController
Please do not apply to your characters without my assistance, it's a complicated process.  This is just what I have so far and it's working out.  You need to apply multiple tags to different items and layers, you need to have a working animator with the right bools, and you need the other code to apply other aspects of this code.  Please take caution when approaching this code and I'd recommend getting me to apply it so we don't mess anything up.