Destroy Self script
This will delete any item applied to after a set amount of time.  The time set is up to you, so it can be anything.  I find it useful for projectiles, so that if they travel straight or have a large area to fly, they won't randomly hit an enemy or a switch and break something.